DuplicateDetector/requirements.txt
Flask==3.1.0
Pillow==11.1.0
imagehash==4.3.2
Flask-CORS==4.0.0
Werkzeug==3.1.3
